# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server for sorting wrong problems.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "1.116.126.208:8080",
    "paths": {
...
        "/app/admin/get_student": {
            "get": {
                "description": "管理员获取一个学生信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "管理员获取一个学生信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "学生id",
                        "name": "Sid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Controllers.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/Services.Student"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/app/admin/get_teacher": {
            "get": {
                "description": "管理员获取一个教师信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "管理员获取一个教师信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "教师id",
                        "name": "Tid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Controllers.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/Services.Teacher"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/app/admin/insert_student": {
            "post": {
                "description": "管理员增加一个学生信息，注册学生",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "管理员增加学生",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "cid",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "cname",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "name": "sgrade",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "sgradeName",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "sicon",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "name": "sid",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "sname",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "snickname",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "soldPassword",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "spassword",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "spasswordAgain",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "sphone",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Controllers.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/Services.Student"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/app/admin/insert_teacher": {
            "post": {
                "description": "管理员增加一个教师信息，注册教师",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "管理员增加教师",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "name": "cnames",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "ticon",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "name": "tid",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "tname",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "tnickname",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "toldPassword",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "tpassword",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "tpasswordAgain",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "tphone",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Controllers.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/Services.Teacher"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/app/admin/update_student": {
            "post": {
                "description": "管理员编辑一个学生信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "管理员编辑学生",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "cid",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "cname",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "name": "sgrade",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "sgradeName",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "sicon",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "name": "sid",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "sname",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "snickname",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "soldPassword",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "spassword",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "spasswordAgain",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "sphone",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Controllers.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/Services.Student"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/app/admin/update_teacher": {
            "post": {
                "description": "管理员编辑一个教师信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "管理员编辑教师",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "name": "cnames",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "ticon",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "name": "tid",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "tname",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "tnickname",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "toldPassword",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "tpassword",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "tpasswordAgain",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "tphone",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Controllers.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/Services.Teacher"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/app/campus/delete_campus/:id": {
            "delete": {
                "description": "删除校区数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "删除校区",
                "parameters": [
                    {
                        "type": "string",
                        "description": "校区id",
                        "name": "Cid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\":200,\"msg\":\"Delete() success!\",\"data\":500001}",
                        "schema": {
                            "$ref": "#/definitions/Controllers.Res"
                        }
                    }
                }
            }
        },
...
        "/app/student/delete_student/:id": {
            "delete": {
                "description": "删除学生数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "删除学生",
                "parameters": [
                    {
                        "type": "string",
                        "description": "学生id",
                        "name": "Sid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\":200,\"msg\":\"Delete() success!\",\"data\":500001}",
                        "schema": {
                            "$ref": "#/definitions/Controllers.Res"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Controllers.Res": {
            "type": "object",
            "properties": {
                "Code": {
                    "type": "integer"
                },
                "Data": {
                    "type": "object"
                },
                "Msg": {
                    "type": "string"
                }
            }
        },
...
        "Services.Student": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "integer"
                },
                "cname": {
                    "type": "string"
                },
                "sgrade": {
                    "type": "integer"
                },
                "sgradeName": {
                    "type": "string"
                },
                "sicon": {
                    "type": "string"
                },
                "sid": {
                    "type": "integer"
                },
                "sname": {
                    "type": "string"
                },
                "snickname": {
                    "type": "string"
                },
                "soldPassword": {
                    "type": "string"
                },
                "spassword": {
                    "type": "string"
                },
                "spasswordAgain": {
                    "type": "string"
                },
                "sphone": {
                    "type": "string"
                }
            }
        },
...
    }
}